Make BarPlot component, remove a lot of the common code
