+++ lib/CountyPlots.js
@@ -1,25 +1,14 @@
 import { useRouter } from 'next/router'
-import { fieldsGenerator, makeBarChartSpec } from './plots.js'
+import BarPlot from './BarPlot.js'
 import { useMemo, useCallback } from 'react'
 import { useFetch } from './queries.js'
 import us from 'us'
-import ContainerDimensions from 'react-container-dimensions'
-
-import { VegaLite } from 'react-vega'
 import WindowSelectSearch from './WindowSelectSearch.js'
-
-const fields = Array.from(fieldsGenerator())
-const filterFields = Array.from(fieldsGenerator(['units'], [''], ['']))
 
 const fuseOptions = {
   keys: ['name'],
   threshold: 0.1,
   distance: 5
-}
-
-function spec (width, height) {
-  // TODO add dropdown for units/property value
-  return makeBarChartSpec(fields, 'units', filterFields, width, height, false)
 }
 
 function getStateAbbreviation (stateCode) {
@@ -62,7 +51,6 @@ export default function CountyPlots ({ countyName, stateAbbr, stateCode }) {
 
   const { status, data: countiesList } = useFetch('/counties_list.json')
 
-  console.log(status)
   const [countyOptions, countyLookup] = useMemo(
     () => makeCountyOptions(countiesList ?? []), [status]
   )
@@ -107,13 +95,7 @@ export default function CountyPlots ({ countyName, stateAbbr, stateCode }) {
       </div>
 
       <div className='w-full flex flex-row'>
-        <ContainerDimensions>
-          {
-        ({ width, height }) => (
-          <VegaLite spec={spec(width, height)} data={{ table: data }} />
-        )
-      }
-        </ContainerDimensions>
+        <BarPlot data={{ table: data }} units='units' perCapita={false} />
       </div>
     </div>
   )
