+++ lib/MetroPlots.js
@@ -1,19 +1,8 @@
 import { useRouter } from 'next/router'
-import { fieldsGenerator, makeBarChartSpec } from './plots.js'
+import BarPlot from './BarPlot.js'
 import { useMemo, useEffect, useCallback } from 'react'
 import { useFetch } from './queries.js'
-
-import { VegaLite } from 'react-vega'
 import WindowSelectSearch from './WindowSelectSearch.js'
-import ContainerDimensions from 'react-container-dimensions'
-
-const fields = Array.from(fieldsGenerator())
-const filterFields = Array.from(fieldsGenerator(['units'], [''], ['']))
-
-function spec (width, height) {
-  // TODO: Add buildings/property value here
-  return makeBarChartSpec(fields, 'units', filterFields, width, height, false)
-}
 
 function getJsonUrl (metro) {
   if (metro === null || (typeof metro === 'undefined')) {
@@ -132,6 +121,7 @@ export default function MetroPlots ({ metroPath, setTitle }) {
   )
   /* eslint-enable */
 
+  // TODO: Add buildings/property option
   return (
     <div className='mx-auto mb-10 align-center items-center flex flex-col justify-center'>
       <div className='lg:grid lg:grid-cols-5 flex flex-col'>
@@ -151,13 +141,7 @@ export default function MetroPlots ({ metroPath, setTitle }) {
       </div>
 
       <div className='w-full flex flex-row'>
-        <ContainerDimensions>
-          {
-          ({ width, height }) => (
-            <VegaLite spec={spec(width, height)} data={{ table: data }} />
-          )
-        }
-        </ContainerDimensions>
+        <BarPlot data={{ table: data }} units='units' perCapita={false} />
       </div>
       {countyList}
     </div>
